Preserve ordering of files for transcoding
Use of an unordered set for capturing the file list caused transcode jobs
to be executed in no particular order.

This commit preserves file order supplied on the command line with depth
first walking of directories.  Full collections in directories will now be
completed before starting on a new directory.